tomo - fix comments
